Made edits according to requested changes (major: removed device setting as it did nothing, moved everything related to i/o into separate file)
